Complete the copy and rename for clarity
