+++ conf/riff-raff.yaml
@@ -48,11 +48,12 @@ deployments:
       functionNames:
         - "media-atom-pipeline-GetChunkFromS3-"
         - "media-atom-pipeline-UploadChunkToYouTube-"
-        - "media-atom-pipeline-CreateCompleteVideoInS3-"
+        - "media-atom-pipeline-MultipartCopyChunkInS3-"
         - "media-atom-pipeline-SendToPluto-"
         - "media-atom-pipeline-SendToTranscoder-"
         - "media-atom-pipeline-GetTranscodingProgress-"
         - "media-atom-pipeline-AddAssetToAtom-"
         - "media-atom-pipeline-AddUploadDataToCache-"
+        - "media-atom-pipeline-CompleteMultipartCopy-"
     dependencies:
       - media-atom-pipeline-cloudformation
