+++ uploader/src/main/resources/cfn-template.yaml
@@ -166,9 +166,9 @@ Resources:
 
   {{UploadChunkToYouTube}}
 
-  {{CreateCompleteVideoInS3}}
+  {{MultipartCopyChunkInS3}}
 
-  {{DeletePartsFromS3}}
+  {{CompleteMultipartCopy}}
 
   {{SendToPluto}}
